refactor reviewed code
