--- packages/react/src/menu-trigger/MenuTrigger.css.ts
@@ -1,13 +1,0 @@
-import { recipe } from "../vanilla-extract";
-
-export const base = recipe({
-  base: [],
-});
-
-export const triggerWrap = recipe({
-  base: [],
-});
-
-export const trigger = recipe({
-  base: [],
-});
