+++ packages/web-components/package.json
@@ -36,10 +36,10 @@
     "./Input": "./dist/components/Input.js",
     "./Kbd": "./dist/components/Kbd.js",
     "./Link": "./dist/components/Link.js",
-    "./Menu": "./dist/Menu.js",
-    "./MenuContent": "./dist/MenuContent.js",
-    "./MenuItem": "./dist/MenuItem.js",
-    "./MenuTrigger": "./dist/MenuTrigger.js",
+    "./Menu": "./dist/components/Menu.js",
+    "./MenuContent": "./dist/components/MenuContent.js",
+    "./MenuItem": "./dist/components/MenuItem.js",
+    "./MenuTrigger": "./dist/components/MenuTrigger.js",
     "./Paper": "./dist/components/Paper.js",
     "./Progress": "./dist/components/Progress.js",
     "./RadioGroup": "./dist/components/RadioGroup.js",
