+++ packages/web-components/src/components/MenuContent.ts
@@ -2,9 +2,9 @@ import { MenuContent as MenuContentComponent } from "@optiaxiom/react";
 
 import type { ComponentAttributes } from "./ComponentAttributes";
 
-import { register } from "./register";
+import { register } from "../register";
 
-export const MenuContent = "ax-dropdown-content";
+export const MenuContent = "ax-menu-content";
 register(MenuContent, MenuContentComponent);
 
 declare global {
