+++ packages/web-components/src/components/MenuTrigger.ts
@@ -2,9 +2,9 @@ import { MenuTrigger as MenuTriggerComponent } from "@optiaxiom/react";
 
 import type { ComponentAttributes } from "./ComponentAttributes";
 
-import { register } from "./register";
+import { register } from "../register";
 
-export const MenuTrigger = "ax-dropdown-trigger";
+export const MenuTrigger = "ax-menu-trigger";
 register(MenuTrigger, MenuTriggerComponent);
 
 declare global {
